+++ src/main/java/com/ghm/giftcardfleamarket/brand/controller/BrandController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ghm.giftcardfleamarket.brand.dto.BrandListResponse;
+import com.ghm.giftcardfleamarket.brand.dto.response.BrandListResponse;
 import com.ghm.giftcardfleamarket.brand.service.BrandService;
 
 import lombok.RequiredArgsConstructor;
